fix(chat): add agent and toolset not available message for public quick app (Issue #4780)
